put npm download count into own component
