+++ src/components/layout/NpmDownloadCount.stories.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import NpmDownloadCount from './NpmDownloadCount';
+
+storiesOf('layout/NpmDownloadCount', module).add('default', () => <NpmDownloadCount />);
